Version 3.4.4 It looks great, i guess... NEW: -4 directions to move. SOON: -Another way to shoot. -The nave follows the pointer (disable)

+++ src/Model/Enemy.java
@@ -33,6 +33,6 @@ public class Enemy extends Sprite {
         this.posY += VELOCIDAD;
     }
     public void dibujar(Graphics g){
-        g.fillOval(this.posX, this.posY, this.width, this.height);
+        g.fillOval((int)this.posX,(int) this.posY, this.width, this.height);
     }
 }
